refactor: simplified the code, got rif of unnecessary interface abuse in order to simply the code, but violated open-closed principle, any new function on data needs to be added to existing class rather then implement interface
