+++ src/main/java/com/codecool/sheetSQL/service/DataFunctionApplyer.java
@@ -1,6 +1,5 @@
 package com.codecool.sheetSQL.service;
 
-import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -8,8 +7,11 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 @Service
-public class ParticularColumnsReaderImpl implements ParticularColumnsReader, DataProcessingHelper {
-
+public class DataFunctionApplyer implements Procesable, DataProcessingHelper {
+    @Override
+    public List<String> getWholeTable(List<String> data) {
+        return new ArrayList<>(data);
+    }
 
     @Override
     public List<String> getChosenColumnsFromRow(List<String> data, List<String> chosenColumns) {
@@ -28,4 +30,20 @@ public class ParticularColumnsReaderImpl implements ParticularColumnsReader, Dat
 
         return result;
     }
+
+    @Override
+    public List<String> getAllColumnsWhere(List<String> data, String columnName, String valueToFind) {
+        String[] headers = getTableHeaders(data).split(",");
+        int columnIndex = findColumnIndex(headers, columnName);
+
+        List<String> resultList =
+                data.stream()
+                        .map(row -> row.split(","))
+                        .filter(array -> array[columnIndex].equals(valueToFind))
+                        .map(array -> String.join(" ", array))
+                        .collect(Collectors.toList());
+        return resultList;
+    }
+
+
 }
