--- src/main/java/com/codecool/sheetSQL/service/Readable.java
@@ -1,8 +1,0 @@
-package com.codecool.sheetSQL.service;
-
-import java.util.List;
-
-public interface Readable {
-
-     List<String> process(List<String> data);
-}
