--- src/main/java/com/codecool/sheetSQL/service/WholeTableReader.java
@@ -1,10 +1,0 @@
-package com.codecool.sheetSQL.service;
-
-import java.util.List;
-
-public interface WholeTableReader extends Readable {
-
-    List<String> process(List<String> data);
-
-
-}
